REFACTOR odklad/1-prijmy
