+++ __tests__/odklad-prijmy-zo-zahranicia.test.ts
@@ -1,4 +1,4 @@
-import { validate } from '../src/pages/odklad/prijmy-zo-zahranicia'
+import { validate } from '../src/plugins/odklad/1-prijmy/prijmy-zo-zahranicia'
 import { testValidation } from './utils/testValidation'
 
 describe('odklad/prijmy-zo-zahranicia', () => {
